sv30_pipeline/requirements.txt
# SV30 Test System - Python Dependencies (Updated for Video-Based Workflow)

# Computer Vision
opencv-python==4.8.1.78
numpy==1.24.3

# Image Processing
Pillow==10.0.0
rembg==2.0.50

# Plotting
matplotlib==3.7.2

# Modbus
pymodbus==3.5.2

# AWS S3 Upload
boto3==1.28.85

# Socket.IO (Real-time Dashboard)
python-socketio==5.10.0

# Optional (for future features)
# Socket.IO client (for dashboard communication)
# python-socketio==5.9.0

# Development/Testing
# pytest==7.4.0
# pytest-cov==4.1.0

# System Dependencies (install via apt):
# sudo apt install ffmpeg  # Required for video recording
